+++ hwad_bench/config/experiment.0.yaml
@@ -8,11 +8,11 @@ train_annotation_path: /app/hwad_bench/store/train.csv
 name: convnext_tiny
 fold: 0
 num_split: 5
-version: v15
+version: v14
 
 embedding_size: 512
 pretrained: true
-validate_score: false
+validate_score: true
 
 use_amp: true
 epochs: 999
