+++ hwad_bench/hwad_bench/data.py
@@ -460,101 +460,3 @@ def add_new_individual(
         new_sub["individual_ids"] = individua_ids[: len(sub["individual_ids"])]
         new_submissions.append(new_sub)
     return new_submissions
-
-
-def search_threshold(
-    train_rows: list[Annotation],
-    val_rows: list[Annotation],
-    submissions: list[Submission],
-    thresholds: list[float],
-) -> list[dict]:
-    train_individual_ids = pipe(
-        train_rows,
-        map(lambda x: x["individual_id"]),
-        set,
-    )
-    val_individual_ids = pipe(
-        val_rows,
-        map(lambda x: x["individual_id"]),
-        set,
-    )
-    print(len(val_rows), len(train_rows))
-    new_individual_ids = val_individual_ids - train_individual_ids
-
-    val_rows = pipe(
-        val_rows,
-        map(
-            lambda x: {
-                **x,
-                "individual_id": "new_individual"
-                if x["individual_id"] in new_individual_ids
-                else x["individual_id"],
-            }
-        ),
-        list,
-    )
-
-    count = 0
-    val_annot_map = pipe(
-        val_rows,
-        map(
-            lambda x: (x["image_file"], x),
-        ),
-        dict,
-    )
-    all_individual_ids = train_individual_ids | set(["new_individual"])
-    label_map = pipe(
-        all_individual_ids,
-        sorted,
-        enumerate,
-        map(lambda x: (x[1], x[0])),
-        dict,
-    )
-    results = []
-    for thr in tqdm(thresholds):
-        metric = MeanAveragePrecisionK()
-        thr_submissions = add_new_individual(submissions, thr)
-        for sub in thr_submissions:
-            labels_at_k = (
-                pipe(
-                    sub["individual_ids"],
-                    map(lambda x: label_map[x]),
-                    list,
-                    torch.tensor,
-                )
-                .long()
-                .unsqueeze(0)
-            )
-            annot = val_annot_map[sub["image_file"]]
-            labels = pipe(
-                [annot["individual_id"]],
-                map(lambda x: label_map[x]),
-                list,
-                torch.tensor,
-            ).long()
-            metric.update(labels_at_k, labels)
-        score, _ = metric.value
-        results.append(
-            {
-                "threshold": thr,
-                "score": score,
-            }
-        )
-    return results
-
-
-def save_submission(
-    submissions: list[Submission],
-    output_path: str,
-) -> list[dict]:
-    rows: list[dict] = []
-    for sub in submissions:
-        image_id = sub["image_file"].split(".")[0]
-        row = {
-            "image": f"{image_id}.jpg",
-            "predictions": " ".join(sub["individual_ids"]),
-        }
-        rows.append(row)
-    df = pd.DataFrame(rows)
-    df.to_csv(output_path, index=False)
-    return rows
