+++ hwad_bench/tests/test_data.py
@@ -14,8 +14,6 @@ from hwad_bench.data import (
     merge_to_coco_rows,
     read_csv,
     read_rows,
-    save_submission,
-    search_threshold,
 )
 from hwad_bench.models import get_writer
 from vision_tools.utils import load_config
@@ -166,52 +164,6 @@ def test_filter_rows_by_fold() -> None:
     assert filtered[0]["image_file"] == "img0-box0.png"
 
 
-def test_search_thresold() -> None:
-    val_rows: list[Any] = [
-        {
-            "image_file": "img0-box0.png",
-            "species": "species-0",
-            "individual_id": "val-0",
-            "label": 1,
-        },
-        {
-            "image_file": "img1--b0.png",
-            "species": "species-0",
-            "individual_id": "tr-0",
-            "label": 2,
-        },
-    ]
-    train_rows: list[Any] = [
-        {
-            "image_file": "img0-box0.png",
-            "species": "species-0",
-            "individual_id": "tr-0",
-            "label": 1,
-        },
-        {
-            "image_file": "img1--b0.png",
-            "species": "species-0",
-            "individual_id": "tr-1",
-            "label": 2,
-        },
-    ]
-    submissions: list[Any] = [
-        dict(
-            image_file="img0-box0.png",
-            distances=[0.9, 0.8, 0.7, 0.6, 0.5],
-            individual_ids=["tr-0", "tr-1", "tr-1", "tr-1", "tr-1"],
-        ),
-    ]
-
-    res = search_threshold(
-        train_rows=train_rows,
-        val_rows=val_rows,
-        submissions=submissions,
-        thresholds=[0.5, 0.95],
-    )
-    assert res == [{"threshold": 0.5, "score": 0.0}, {"threshold": 0.95, "score": 1.0}]
-
-
 def test_add_new_individual() -> None:
     submissions: list[Any] = [
         dict(
@@ -227,15 +179,3 @@ def test_add_new_individual() -> None:
     )
     assert len(res) == 1
     assert res[0]["individual_ids"][3] == "new_individual"
-
-
-def test_save_submission() -> None:
-    submissions: list[Any] = [
-        dict(
-            image_file="img0-xxxxx.jpg",
-            individual_ids=["tr-0", "tr-1", "tr-1", "tr-1", "tr-1"],
-        )
-    ]
-    rows = save_submission(submissions, "/app/test_outputs/submissions.csv")
-    assert len(rows) == 1
-    assert rows[0]["image"] == "img0.jpg"
